requirements.txt
streamlit
requests
beautifulsoup4
sentence-transformers
scikit-learn
numpy
python-dotenv
google-generativeai
newsapi-python
